Move src/imap_client.rs into src/imap/client.rs
